Fix PSScriptAnalyzer results

+++ JiraPS/Public/Remove-JiraVersion.ps1
@@ -1,5 +1,4 @@
-﻿function Remove-JiraVersion
-{
+﻿function Remove-JiraVersion {
     <#
     .Synopsis
        This function removes an existing version.
@@ -18,32 +17,35 @@
     .NOTES
        This function requires either the -Credential parameter to be passed or a persistent JIRA session. See New-JiraSession for more details.  If neither are supplied, this function will run with anonymous access to JIRA.
     #>
-    [CmdletBinding(DefaultParameterSetName = 'VersionID')]
+    [CmdletBinding(
+        SupportsShouldProcess = $true,
+        DefaultParameterSetName = 'VersionID'
+    )]
     param(
         # Jira Version Name
         [Parameter(Mandatory = $true,
-                    ValueFromPipelineByPropertyName=$true,
-                    ParameterSetName = 'Key')]
+            ValueFromPipelineByPropertyName = $true,
+            ParameterSetName = 'Key')]
         [Alias('Versions')]
         [string] $Name,
 
         # The Project ID or project key of a project to search
         [Parameter(Mandatory = $true,
-                    ValueFromPipelineByPropertyName=$true,
-                    ParameterSetName = 'Key',
-                    Position = 0,
-                    ValueFromRemainingArguments = $true)]
+            ValueFromPipelineByPropertyName = $true,
+            ParameterSetName = 'Key',
+            Position = 0,
+            ValueFromRemainingArguments = $true)]
         [String] $Project,
 
         # The Version ID
         [Parameter(Mandatory = $true,
-                    ValueFromPipelineByPropertyName=$true,
-                    ParameterSetName = 'VersionID')]
+            ValueFromPipelineByPropertyName = $true,
+            ParameterSetName = 'VersionID')]
         [String] $ID,
 
         # The Version URL
         [Parameter(Mandatory = $false,
-                    ValueFromPipelineByPropertyName=$true)]
+            ValueFromPipelineByPropertyName = $true)]
         [Alias('Self')]
         [String] $RestURL,
 
@@ -52,16 +54,13 @@
         [System.Management.Automation.PSCredential] $Credential
     )
 
-    begin
-    {
+    begin {
         Write-Debug -Message '[Remove-JiraVersion] Reading information from config file'
-        try
-        {
+        try {
             Write-Debug -Message '[Remove-JiraVersion] Reading Jira server from config file'
             $server = Get-JiraConfigServer -ConfigFile $ConfigFile -ErrorAction Stop
         }
-        catch
-        {
+        catch {
             $err = $_
             Write-Debug -Message '[Remove-JiraVersion] Encountered an error reading configuration data.'
             throw $err
@@ -69,36 +68,31 @@
 
         Write-Debug "[Remove-JiraVersion] Completed Begin block."
     }
-    process
-    {
-        Switch($PSCmdlet.ParameterSetName)
-        {
-            'Project'
-            {
+    process {
+        Switch ($PSCmdlet.ParameterSetName) {
+            'Project' {
                 $existingVersion = Get-JiraVersion -Project $Project | Where-Object {$PSItem.Name -eq $Name}
                 $restUrl = $existingVersion.self
             }
-            'VersionID'
-            {
+            'VersionID' {
                 $restUrl = "$server/rest/api/2/version/$ID"
             }
         }
         Write-Debug "[Get-JiraVersion] Rest URL set to $restUrl."
 
-        Write-Debug -Message '[Remove-JiraVersion] Preparing for blastoff!'
-        $result = Invoke-JiraMethod -Method Delete -URI $restUrl -Credential $Credential
+        if ($PSCmdlet.ShouldProcess(($RestURL.replace($server, "")), "Removing Version on JIRA")) {
+            Write-Debug -Message '[Remove-JiraVersion] Preparing for blastoff!'
+            $result = Invoke-JiraMethod -Method Delete -URI $restUrl -Credential $Credential
+        }
 
-        If ($result)
-        {
+        If ($result) {
             Write-Output -InputObject $result
         }
-        Else
-        {
+        Else {
             Write-Debug -Message '[Remove-JiraVersion] Jira returned no results to output.'
         }
     }
-    end
-    {
+    end {
         Write-Debug "[Remove-JiraVersion] Complete"
     }
 }
